+++ cylc/flow/cfgspec/globalcfg.py
@@ -984,7 +984,9 @@ with Conf('global.cylc', desc='''
 
             .. versionadded:: 8.0.0
         """):
-            with Conf('<install target>'):
+            with Conf('<install target>', desc="""
+                :ref:`Host <Install targets>` on which to create the symlinks.
+            """):
                 Conf('run', VDR.V_STRING, None, desc="""
                     Alternative location for the run dir.
 
@@ -1170,8 +1172,7 @@ with Conf('global.cylc', desc='''
                 .. versionadded:: 8.0.0
             ''')
             Conf('job runner', VDR.V_STRING, 'background', desc=f'''
-                The batch system/job submit method used to run jobs on the
-                platform.
+                The system used to run jobs on the platform.
 
                 Examples:
 
@@ -1287,8 +1288,10 @@ with Conf('global.cylc', desc='''
                  VDR.V_STRING,
                  'rsync',
                  desc='''
-                Command used for remote file installation. This supports POSIX
-                compliant rsync implementation e.g. GNU or BSD.
+                Command used for file installation.
+
+                This supports POSIX compliant rsync implementations e.g. GNU or
+                BSD.
 
                 .. versionadded:: 8.0.0
             ''')
@@ -1640,9 +1643,25 @@ with Conf('global.cylc', desc='''
 
             .. versionadded:: 8.0.0
         '''):
-            Conf('hosts', VDR.V_STRING_LIST, ['localhost'])
-            with Conf('selection', meta=Selection):
-                Conf('method', VDR.V_STRING, default='definition order')
+            Conf('hosts', VDR.V_STRING_LIST, ['localhost'], desc='''
+                .. seealso::
+
+                   :cylc:conf:`global.cylc[platforms][<platform name>]hosts`
+            ''')
+            with Conf(
+                'selection', meta=Selection,
+                desc=(
+                    ':cylc:conf:`global.cylc[platforms][<platform name>]'
+                    '[selection]`'
+                )
+            ):
+                Conf('method', VDR.V_STRING, default='definition order',
+                     desc='''
+                        .. seealso::
+
+                           :cylc:conf:`global.cylc[platforms][<platform name>]
+                           [selection]method`
+                ''')
 
     # Platform Groups
     with Conf('platform groups', desc='''
@@ -1667,7 +1686,10 @@ with Conf('global.cylc', desc='''
 
         .. versionadded:: 8.0.0
     '''):  # noqa: SIM117 (keep same format)
-        with Conf('<group>'):
+        with Conf('<group>', desc='''
+        The name of a :cylc:conf:`platform group
+        <global.cylc[platform groups]>`.
+        '''):
             with Conf('meta', desc=PLATFORM_META_DESCR):
                 Conf('<custom metadata>', VDR.V_STRING, '', desc='''
                     Any user-defined metadata item.
@@ -1688,7 +1710,10 @@ with Conf('global.cylc', desc='''
                    host.
 
             ''')
-            with Conf('selection'):
+            with Conf(
+                'selection',
+                desc='Sets how platforms are selected from platform groups.'
+            ):
                 Conf(
                     'method', VDR.V_STRING, default='random',
                     options=['random', 'definition order'],