+++ cylc/flow/scripts/common.py
@@ -16,7 +16,6 @@
 
 """Shared utilities for Cylc scripts."""
 
-from re import sub
 from itertools import zip_longest
 from ansimarkup import strip
 
@@ -24,42 +23,13 @@ from cylc.flow import __version__
 from cylc.flow.scripts import _copyright_year
 from cylc.flow.terminal import get_width
 
-
 # fmt: off
-LOGO_LETTERS = (
-    (
-        "ooo",
-        "oo ",
-        "oooo",
-    ),
-    (
-        "oo oo",
-        "ooooo",
-        " ooo",
-    ),
-    (
-        "oo ",
-        "ooo",
-        "ooo",
-    ),
-    (
-        "oooo",
-        "oo  ",
-        "oooo",
-    )
-)
-# fmt: on
-
-LOGO = [
-    ''.join(
-        sub('o', f'<white,{tag}> </white,{tag}>', letter[ind])
-        for tag, letter in zip(
-            ('red', 'yellow', 'green', 'blue'),
-            LOGO_LETTERS
-        )
-    )
-    for ind in range(len(LOGO_LETTERS[0]))
+CYLC_LOGO = [
+    " <yellow>▪</yellow> <blue>■ </blue>",
+    " <green>██</green>  ",
+    "<red>▝▘</red>   ",
 ]
+# fmt: on
 
 CYLC = (
     "<b>"
@@ -84,16 +54,16 @@ def cylc_header(width=None):
     width = width or get_width()
     lmax = (
         max(len(strip(line)) for line in LICENSE) +
-        len(strip(LOGO[0]))
+        len(strip(CYLC_LOGO[0]))
     )
     if width >= lmax + 1:
         header = '\n'.join(
             ('{0} {1}').format(*x)
             for x in zip_longest(
-                LOGO,
+                CYLC_LOGO,
                 LICENSE
             )
         )
     else:
-        header = '\n'.join(LOGO) + '\n' + '\n'.join(LICENSE)
+        header = '\n'.join(CYLC_LOGO) + '\n' + '\n'.join(LICENSE)
     return f"\n{header}\n"