+++ cylc/flow/task_job_mgr.py
@@ -36,6 +36,7 @@ from logging import (
 )
 from shutil import rmtree
 from time import time
+from typing import TYPE_CHECKING
 
 from cylc.flow import LOG
 from cylc.flow.job_runner_mgr import JobPollContext
@@ -111,6 +112,9 @@ from cylc.flow.wallclock import (
     get_utc_mode
 )
 from cylc.flow.cfgspec.globalcfg import SYSPATH
+
+if TYPE_CHECKING:
+    from cylc.flow.task_proxy import TaskProxy
 
 
 class TaskJobManager:
@@ -1079,7 +1083,12 @@ class TaskJobManager:
                 itask, CRITICAL, self.task_events_mgr.EVENT_SUBMIT_FAILED,
                 ctx.timestamp)
 
-    def _prep_submit_task_job(self, workflow, itask, check_syntax=True):
+    def _prep_submit_task_job(
+        self,
+        workflow: str,
+        itask: 'TaskProxy',
+        check_syntax: bool = True
+    ):
         """Prepare a task job submission.
 
         Returns:
@@ -1171,7 +1180,9 @@ class TaskJobManager:
                 rtconfig['remote']['host'] = host_n
 
             try:
-                platform = get_platform(rtconfig, bad_hosts=self.bad_hosts)
+                platform = get_platform(
+                    rtconfig, itask.tdef.name, bad_hosts=self.bad_hosts
+                )
 
             except PlatformLookupError as exc:
                 itask.waiting_on_job_prep = False