+++ tests/unit/test_job_file.py
@@ -118,8 +118,8 @@ def test_write(fixture_get_platform):
 
     expected = ('#!/bin/bash -l\n#\n# ++++ THIS IS A CYLC TASK JOB SCRIPT '
                 '++++\n# Workflow: farm_noises\n# Task: 1/baa\n# Job '
-                'log directory: 1/baa/01\n# Job submit method: '
-                'background\n# Job submit command template: woof\n#'
+                'log directory: 1/baa/01\n# Job runner: '
+                'background\n# Job runner command template: woof\n#'
                 ' Execution time limit: moo')
 
     platform = fixture_get_platform(